practice for section 4.5
Description：Given a pair of positive integers, for example, 6 and 110, can this equation 6-110 be true? The answer is"yes", if 6 is a decimal number and 110 is a binary number Now for any pair of positive integers NI and N2, your task is to find the radix of one number while that of the other is given 
Input：Each input file contains one test case. Each case occupies a line which contains 4 positive Integers NI N2 tag radix Here N1 and N2 each has no more than 10 digits. A digit is less than its radix and is chosen from the set (0-9, a-z where 0-9 represent the decimal numbers 0-9, and a-z represent the decimal numbers 10-35. The last number"radix"is the radix of NI if"tag"is 1, or of N2 if"tag"is 2 
Output：For each test case, print in one line the radix of the other number so that the equation NI= N2 is true. If the equation is impossible, print"Impossible". If the solution is not unique, output the smallest possible radix
Input Sample：1 6 110 1 10 
Output Sample：2
Input Sample： 1 ab 1 2 
Output Sample： Impossible